// src/components/uniswap/src/types/position.d.ts
import { BigNumber } from '@ethersproject/bignumber'

export interface PositionDetails {
  nonce: BigNumber
  tokenId: BigNumber
  operator: string
  token0: string
  token1: string
  fee: number
  tickLower: number
  tickUpper: number
  liquidity: BigNumber
  feeGrowthInside0LastX128: BigNumber
  feeGrowthInside1LastX128: BigNumber
  slot0: {
    liquidity: number,
    sqrtPriceX96: BigNumber,
    tick: number
  },
  tokensOwed0: BigNumber
  tokensOwed1: BigNumber
}